Consolidate refactored tape downloader, add tests.
