+++ test/test_GD.py
@@ -19,6 +19,19 @@ class TestGD(unittest.TestCase):
             # slow: initializes a new archive synchronously
             GD.GDArchive(directory, sync=True)
 
+    def _test_tape_downloader(self, downloader):
+        tapes = downloader.get_tapes([1970])
+        frac_id = len([x for x in tapes if "identifier" in x]) / len(tapes)
+        self.assertGreaterEqual(len(tapes), 200, msg="1970 has many tapes")        
+        self.assertEquals(frac_id, 1.0, msg="All tapes have an identifier")
+
+    def test_sync_tape_downloader(self):
+        downloader = GD.TapeDownloader()
+        self._test_tape_downloader(downloader)
+    
+    def test_async_tape_downloader(self):
+        downloader = GD.AsyncTapeDownloader()
+        self._test_tape_downloader(downloader)
 
 if __name__ == '__main__':
     unittest.main()
